prevent compile error and set platform to iOS

+++ UIImage+Resize.m
@@ -92,7 +92,8 @@
             break;
             
         default:
-            [NSException raise:NSInvalidArgumentException format:@"Unsupported content mode: %d", contentMode];
+            [NSException raise:NSInvalidArgumentException 
+                        format:@"Unsupported content mode: %lu", (unsigned long)contentMode];
     }
     
     CGSize newSize = CGSizeMake(self.size.width * ratio, self.size.height * ratio);
